🛠️: refactor `FormHidden` test to typescript
